+++ umami/tests/unit/preprocessing/test_preprocessing_tools_resampling.py
@@ -1,5 +1,6 @@
 """Unit tests for Resampling in the preprocessing tools."""
 import os
+import shutil
 import tempfile
 import unittest
 from subprocess import run
@@ -9,7 +10,8 @@ import numpy as np
 import pandas as pd
 
 from umami.configuration import global_config, logger, set_log_level
-from umami.preprocessing_tools import (  # PDFSampling,
+from umami.preprocessing_tools import (
+    PDFSampling,
     PreprocessConfiguration,
     UnderSampling,
     UnderSamplingNoReplace,
@@ -253,181 +255,17 @@ class ResamplingTestCase(unittest.TestCase):
     # TODO: write tests
 
 
-class UnderSamplingTestCase(unittest.TestCase):
+class ResamplingTestCaseComplexDistr(unittest.TestCase):
     """
-    Test the implementation of the UnderSampling class.
-    """
-
-    def setUp(self):
-        """
-        Create a default dataset for testing.
-        """
-        self.config_file = os.path.join(
-            os.path.dirname(__file__), "fixtures", "test_preprocess_config.yaml"
-        )
-        self.config = PreprocessConfiguration(self.config_file)
-        self.sampling_config = self.config.sampling
-
-        self.df_bjets = pd.DataFrame(
-            {
-                global_config.pTvariable: abs(np.random.normal(300000, 30000, 10000)),
-                global_config.etavariable: abs(np.random.normal(1.25, 1, 10000)),
-            }
-        )
-        self.df_cjets = pd.DataFrame(
-            {
-                global_config.pTvariable: abs(np.random.normal(280000, 28000, 10000)),
-                global_config.etavariable: abs(np.random.normal(1.4, 1, 10000)),
-            }
-        )
-        self.df_ujets = pd.DataFrame(
-            {
-                global_config.pTvariable: abs(np.random.normal(250000, 25000, 10000)),
-                global_config.etavariable: abs(np.random.normal(1.0, 1, 10000)),
-            }
-        )
-
-    def test_count_no_samples_defined(self):
-        """Test no samples defined."""
-        del self.sampling_config.options.samples_training
-        us_norepl = UnderSampling(self.config)
-        with self.assertRaises(KeyError):
-            us_norepl.initialise_samples()
-
-    def test_different_samples_per_category(self):
-        """Test different samples per category."""
-        del self.sampling_config.options.samples_training["zprime"][1]
-        us_norepl = UnderSampling(self.config)
-        with self.assertRaises(RuntimeError):
-            us_norepl.initialise_samples()
-
-
-# TODO: this can be used to extend the UnderSamplingTestCase
-# class UnderSamplingOldTestCase(unittest.TestCase):
-#     """
-#     Test the implementation of the UnderSampling class.
-#     """
-
-#     def setUp(self):
-#         """
-#         Create a default dataset for testing.
-#         """
-#         self.df_bjets = pd.DataFrame(
-#             {
-#                 global_config.pTvariable: abs(
-#                     np.random.normal(300000, 30000, 10000)
-#                 ),
-#                 global_config.etavariable: abs(
-#                     np.random.normal(1.25, 1, 10000)
-#                 ),
-#             }
-#         )
-#         self.df_cjets = pd.DataFrame(
-#             {
-#                 global_config.pTvariable: abs(
-#                     np.random.normal(280000, 28000, 10000)
-#                 ),
-#                 global_config.etavariable: abs(
-#                     np.random.normal(1.4, 1, 10000)
-#                 ),
-#             }
-#         )
-#         self.df_ujets = pd.DataFrame(
-#             {
-#                 global_config.pTvariable: abs(
-#                     np.random.normal(250000, 25000, 10000)
-#                 ),
-#                 global_config.etavariable: abs(
-#                     np.random.normal(1.0, 1, 10000)
-#                 ),
-#             }
-#         )
-
-#     def test_zero_case(self):
-#         df_zeros = pd.DataFrame(
-#             np.zeros((1000, 2)),
-#             columns=[global_config.pTvariable, global_config.etavariable],
-#         )
-#         down_s = UnderSampling(df_zeros, df_zeros, df_zeros)
-#         b_ind, c_ind, u_ind, _ = down_s.GetIndices()
-#         self.assertEqual(len(b_ind), len(df_zeros))
-
-#     def test_underflow(self):
-#         df_minus_ones = pd.DataFrame(
-#             -1 * np.ones((1000, 2)),
-#             columns=[global_config.pTvariable, global_config.etavariable],
-#         )
-#         down_s = UnderSampling(df_minus_ones, df_minus_ones, df_minus_ones)
-#         b_ind, c_ind, u_ind, _ = down_s.GetIndices()
-#         self.assertEqual(b_ind.size, 0)
-#         self.assertEqual(c_ind.size, 0)
-#         self.assertEqual(u_ind.size, 0)
-
-#     def test_overflow(self):
-#         df_large = pd.DataFrame(
-#             1e10 * np.ones((1000, 2)),
-#             columns=[global_config.pTvariable, global_config.etavariable],
-#         )
-#         down_s = UnderSampling(df_large, df_large, df_large)
-#         b_ind, c_ind, u_ind, _ = down_s.GetIndices()
-#         self.assertEqual(b_ind.size, 0)
-#         self.assertEqual(c_ind.size, 0)
-#         self.assertEqual(u_ind.size, 0)
-
-#     def test_equal_length(self):
-#         down_s = UnderSampling(self.df_bjets, self.df_cjets, self.df_ujets)
-#         b_ind, c_ind, u_ind, _ = down_s.GetIndices()
-#         self.assertEqual(len(b_ind), len(c_ind))
-#         self.assertEqual(len(b_ind), len(u_ind))
-
-
-class PDFResamplingTestCase(unittest.TestCase):
-    """
-    Test the implementation of the PDFResampling base class.
+    Class that handle creation of a somewhat coplex dataset of test data
+    useful to be inherited by other test classes
     """
 
-    def setUp(self):
+    def generate_test_data(self):
         """
-        Create a default dataset for testing.
+        Generates a relatively complex dataset of test data
         """
-        self.config_file = os.path.join(
-            os.path.dirname(__file__), "fixtures", "test_preprocess_config.yaml"
-        )
-        self.config = PreprocessConfiguration(self.config_file)
-
-    # TODO: adding tests for PDFSampling class
-
-
-class UnderSamplingNoReplaceTestCase(unittest.TestCase):
-    """
-    Test the implementation of the UnderSamplingNoReplace class.
-    """
-
-    def setUp(self):
-        """
-        Create a default dataset for testing.
-        """
-        self.config_file = os.path.join(
-            os.path.dirname(__file__), "fixtures", "test_preprocess_config.yaml"
-        )
-        self.config = PreprocessConfiguration(self.config_file)
-        sampling_config = self.config.sampling
-        sampling_config.options.target_distribution = "bjets"
-        sampling_config.options.sampling_variables[0][global_config.pTvariable][
-            "bins"
-        ] = [
-            0,
-            15e5,
-            21,
-        ]
-        sampling_config.options.sampling_variables[1][global_config.etavariable][
-            "bins"
-        ] = [
-            0,
-            2.5,
-            2,
-        ]
-        self.sampling_config = sampling_config
+        np.random.seed(42)
         self.data = {
             "training_ttbar_bjets": pd.DataFrame(
                 {
@@ -488,6 +326,9 @@ class UnderSamplingNoReplaceTestCase(unittest.TestCase):
         ]
         for sample in training_ttbar_samples:
             test_h5_file_name = self.config.preparation.get_sample(sample).output_name
+            dir_path = os.path.dirname(test_h5_file_name)
+            if dir_path != "" and not os.path.exists(dir_path):
+                os.makedirs(dir_path)
             with h5py.File(test_h5_file_name, "w") as f_h5:
                 jets = f_h5.create_dataset(
                     "jets",
@@ -505,6 +346,330 @@ class UnderSamplingNoReplaceTestCase(unittest.TestCase):
                 jets[global_config.etavariable] = self.data[sample][
                     global_config.etavariable
                 ]
+
+    def setUp(self):
+        """
+        Create a default dataset for testing.
+        """
+        # Read in the configuration file
+        self.config_file = os.path.join(
+            os.path.dirname(__file__),
+            "fixtures",
+            "test_preprocess_config_complex_sampling.yaml",
+        )
+        self.config = PreprocessConfiguration(self.config_file)
+        self.sampling_config = self.config.sampling
+        self.generate_test_data()
+
+    def tearDown(self):
+        """
+        Delete the workspace folder after the test.
+        """
+        shutil.rmtree("test_preprocessing_workspace")
+        # hardcoding here is intentional so that in case
+        # of modifications to the test config file
+        # we don't delete the wrong folder by accident (e.g. the root folder)
+        # plese modify this line if you change the test_parameters.yaml file
+
+
+class ResamplingTestCaseEasyDistr(unittest.TestCase):
+    """
+    Class that handle creation of a somewhat simple dataset of test data
+    useful to be inherited by other test classes
+    """
+
+    def generate_test_data(self):
+        """
+        Generates a relatively simple dataset of test data
+        """
+        np.random.seed(42)
+        n_events = 100000  # number of events to generate
+        self.data = {
+            "training_ttbar_bjets": pd.DataFrame(
+                {
+                    global_config.pTvariable: abs(
+                        np.random.uniform(0, 6000000, n_events)
+                    ),
+                    global_config.etavariable: abs(np.random.normal(1.25, 1, n_events)),
+                }
+            ),
+            "training_ttbar_cjets": pd.DataFrame(
+                {
+                    global_config.pTvariable: abs(
+                        np.random.uniform(0, 6000000, n_events)
+                    ),
+                    global_config.etavariable: abs(np.random.normal(1.4, 1, n_events)),
+                }
+            ),
+            "training_ttbar_ujets": pd.DataFrame(
+                {
+                    global_config.pTvariable: abs(
+                        np.random.uniform(0, 6000000, n_events)
+                    ),
+                    global_config.etavariable: abs(np.random.normal(1.0, 1, n_events)),
+                }
+            ),
+            "training_zprime_bjets": pd.DataFrame(
+                {
+                    global_config.pTvariable: abs(
+                        np.random.uniform(0, 6000000, n_events)
+                    ),
+                    global_config.etavariable: abs(np.random.normal(1.5, 1, n_events)),
+                }
+            ),
+            "training_zprime_cjets": pd.DataFrame(
+                {
+                    global_config.pTvariable: abs(
+                        np.random.uniform(0, 6000000, n_events)
+                    ),
+                    global_config.etavariable: abs(np.random.normal(1.6, 1, n_events)),
+                }
+            ),
+            "training_zprime_ujets": pd.DataFrame(
+                {
+                    global_config.pTvariable: abs(
+                        np.random.uniform(0, 6000000, n_events)
+                    ),
+                    global_config.etavariable: abs(np.random.normal(1.2, 1, n_events)),
+                }
+            ),
+        }
+        training_ttbar_samples = [
+            "training_ttbar_bjets",
+            "training_ttbar_cjets",
+            "training_ttbar_ujets",
+            "training_zprime_bjets",
+            "training_zprime_cjets",
+            "training_zprime_ujets",
+        ]
+        for sample in training_ttbar_samples:
+            test_h5_file_name = self.config.preparation.get_sample(sample).output_name
+            dir_path = os.path.dirname(test_h5_file_name)
+            if not os.path.exists(dir_path):
+                os.makedirs(dir_path)
+            with h5py.File(test_h5_file_name, "w") as f_h5:
+                jets = f_h5.create_dataset(
+                    "jets",
+                    (n_events),
+                    dtype=np.dtype(
+                        [
+                            (global_config.pTvariable, "f"),
+                            (global_config.etavariable, "f"),
+                        ]
+                    ),
+                )
+                jets[global_config.pTvariable] = self.data[sample][
+                    global_config.pTvariable
+                ]
+                jets[global_config.etavariable] = self.data[sample][
+                    global_config.etavariable
+                ]
+
+    def setUp(self):
+        """
+        Create a default dataset for testing.
+        """
+        # Read in the configuration file
+        self.config_file = os.path.join(
+            os.path.dirname(__file__),
+            "fixtures",
+            "test_preprocess_config_easy_samples.yaml",
+        )
+        self.config = PreprocessConfiguration(self.config_file)
+        self.sampling_config = self.config.sampling
+        self.generate_test_data()
+
+    def tearDown(self):
+        """
+        Delete the workspace folder after the test.
+        """
+        shutil.rmtree("test_preprocessing_workspace")
+        # TODO: uncomment this line before merge
+        # hardcoding here is intentional so that in case of
+        # modifications to the test config file
+        # we don't delete the wrong folder by accident (e.g. the root folder)
+        # plese modify this line if you change the test_parameters.yaml file
+
+
+class UnderSamplingTestCase(ResamplingTestCaseComplexDistr):
+    """
+    Test the implementation of the UnderSampling class.
+    """
+
+    def test_count_no_samples_defined(self):
+        """Test no samples defined."""
+        del self.sampling_config.options.samples_training
+        us_norepl = UnderSampling(self.config)
+        with self.assertRaises(KeyError):
+            us_norepl.initialise_samples()
+
+    def test_different_samples_per_category(self):
+        """Test different samples per category."""
+        del self.sampling_config.options.samples_training["zprime"][1]
+        us_norepl = UnderSampling(self.config)
+        with self.assertRaises(RuntimeError):
+            us_norepl.initialise_samples()
+
+    def test_run(self):
+        """
+        Test the run method. and ensure the deterministic sampling for a seed.
+        Most of the major code changes would brake this test.
+        """
+        countsampler = UnderSampling(self.config)
+        countsampler.Run()
+        indices = countsampler.get_indices()
+        # ensure the deterministic sampling for a seed
+        np.testing.assert_array_equal(
+            indices["training_ttbar_bjets"][:10],
+            np.array([7, 8, 13, 16, 18, 28, 29, 33, 35, 36]),
+        )
+        np.testing.assert_array_equal(
+            indices["training_zprime_ujets"][:10],
+            np.array([18, 19, 23, 60, 78, 82, 119, 155, 200, 204]),
+        )
+
+    # def testDESIRED_equal_length(self):
+    #     """Test equal fractions."""
+    #     # ensure the total number of jets is divisible by the number of categories (3)
+    #     n_jets_initial = self.config.sampling["options"]["n_jets"]
+    #     n_jets_to_plot_initial = self.config.sampling["options"]["n_jets_to_plot"]
+    #     self.config.sampling["options"]["n_jets"] = 3e5
+    #     self.config.sampling["options"]["n_jets_to_plot"] = 3e5
+    #     countsampler = UnderSampling(self.config)
+    #     countsampler.Run()
+    #     indices = countsampler.get_indices()
+    #     n_b = len(indices["training_ttbar_bjets"]) + len(
+    #         indices["training_zprime_bjets"]
+    #     )
+    #     n_c = len(indices["training_ttbar_cjets"]) + len(
+    #         indices["training_zprime_cjets"]
+    #     )
+    #     n_u = len(indices["training_ttbar_ujets"]) + len(
+    #         indices["training_zprime_ujets"]
+    #     )
+    #     self.assertEqual(n_b, n_c)
+    #     self.assertEqual(n_b, n_u)
+    #     self.config.sampling["options"]["n_jets"] = n_jets_initial
+    #     self.config.sampling["options"]["n_jets_to_plot"] = n_jets_to_plot_initial
+
+    def test_equal_length_smalln(self):
+        """Test equal fractions."""
+        # ensure the total number of jets is divisible by
+        # the number of categories (3) and small
+        n_jets_initial = self.config.sampling.options.n_jets
+        n_jets_to_plot_initial = self.config.sampling.options.n_jets_to_plot
+        self.config.sampling.options.n_jets = 900
+        self.config.sampling.options.n_jets_to_plot = 900
+        countsampler = UnderSampling(self.config)
+        countsampler.Run()
+        indices = countsampler.get_indices()
+        n_b = len(indices["training_ttbar_bjets"]) + len(
+            indices["training_zprime_bjets"]
+        )
+        n_c = len(indices["training_ttbar_cjets"]) + len(
+            indices["training_zprime_cjets"]
+        )
+        n_u = len(indices["training_ttbar_ujets"]) + len(
+            indices["training_zprime_ujets"]
+        )
+        self.assertEqual(n_b, n_c)
+        self.assertEqual(n_b, n_u)
+        self.config.sampling.options.n_jets = n_jets_initial
+        self.config.sampling.options.n_jets_to_plot = n_jets_to_plot_initial
+
+
+class UnderSamplingTestCaseEasyDistr(ResamplingTestCaseEasyDistr):
+    """
+    Test the implementation of the UnderSampling class, using a simple example.
+    """
+
+    def test_equal_length(self):
+        """Test equal fractions."""
+        countsampler = UnderSampling(self.config)
+        countsampler.Run()
+        indices = countsampler.get_indices()
+
+        n_b = len(indices["training_ttbar_cjets"])
+        n_c = len(indices["training_ttbar_cjets"])
+        n_u = len(indices["training_ttbar_ujets"])
+        self.assertEqual(n_b, n_c)
+        self.assertEqual(n_b, n_u)
+
+
+class PDFResamplingTestCase(ResamplingTestCaseComplexDistr):
+    """
+    Test the run method. and ensure the deterministic sampling for a seed.
+    Most of the major code changes would brake this test.
+    """
+
+    def setUp(self):
+        """
+        Create a default dataset for testing.
+        """
+        super().setUp()
+
+        # tweek configs so that they wrk for PDF sampling
+        self.sampling_config.options.target_distribution = "bjets"
+        self.sampling_config.options.sampling_variables[0][global_config.pTvariable][
+            "bins"
+        ] = [
+            [0, 2.5e5, 11],
+            [
+                2.6e5,
+                5e5,
+                11,
+            ],
+        ]
+        self.sampling_config.options.sampling_variables[1][global_config.etavariable][
+            "bins"
+        ] = [
+            [0, 1.25, 5],
+            [0, 1.25, 5],
+        ]
+
+    def test_no_samples_defined(self):
+        """Test no samples defined."""
+        del self.sampling_config.options.samples_training
+        self.assertRaises(AttributeError, lambda: PDFSampling(self.config))
+
+    def test_run(self):
+        """Test equal fractions."""
+        pdfsampler = PDFSampling(self.config)
+        pdfsampler.Run()
+        output_name = pdfsampler.config.get_file_name(
+            option="resampled",
+            use_val=pdfsampler.use_validation_samples,
+        )
+        with h5py.File(output_name, "r") as file:
+            jets = np.array(file["jets"]["pt_btagJes"][:10])
+        jets_must = np.array(
+            [
+                264299.53,
+                270432.1,
+                267145.62,
+                271890.97,
+                253087.52,
+                321016.5,
+                183927.34,
+                256265.61,
+                196289.08,
+                259284.94,
+            ]
+        )
+        np.testing.assert_allclose(jets, jets_must)
+
+
+class UnderSamplingNoReplaceTestCase(ResamplingTestCaseComplexDistr):
+    """
+    Test the implementation of the UnderSamplingNoReplace class.
+    """
+
+    def setUp(self):
+        """
+        Create a default dataset for testing.
+        """
+        super().setUp()
+        self.sampling_config.options.target_distribution = "bjets"
 
     def test_no_samples_defined(self):
         """Test no samples defined."""