+++ changelog.md
@@ -8,6 +8,7 @@
 
 ### Latest
 
+- Added and refactored some unit tests for resampling classes, refactored Undersampling and UndersamllingNoReplace classes[!705](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/705)
 - Add support for different labels for each track-like group [!688](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/688)
 
 ### [v0.19](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/tags/0.19) (20.03.2023)
@@ -20,6 +21,7 @@
 - Fixing precision problem in discriminant calculation [!701](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/701)
 - Improve scale dict copy logic [!698](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/698)
 - update LWTNN-conversion documentation [!699](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/699)
+- Enable correct usage of dataclasses in preprocessing config [!696](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/696)
 - Fixing common variable bug in PDF sampling [!685](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/685)
 - Add VR track jet configs and fix bug in pdf sampling for a third sample category [!657](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/657)
 - Enabling additional jet labels integration test [!695](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/695)